Moved article endpoint to separate blueprint.
